fuir: move some abstract from GeneratingFUIR to FUIR
